Created the View Map Function
Got an API key from google maps API and created the view map function. In this update the user is provided with the route which is the nearest to him. And once the User is completed with his destination, then the user can press the button which says confirm Destination. Once that button is pressed user is provided with the route for the next destination